make new controlled piano into ControlledPiano component, and make Piano a lightweight uncontrolled component on top of it
